Fix up file download handling. Move debug logging to level 13.
Implemented a FileListing object for tracking files on the devices. If
the device doesn't send any file data for a while it will be asked for
a listing of files. Need to handle failures.
Fix bug where a single packet would result in an Exception as the check
for the file being complete would only occur after processing the second
packet in a group.