Change mt clients to be synchronous. Add timeouts. Enhance firm update
Implement an outbound message queue and change commands to be
synchronous. This could block a single connection due to some messages
being ignored so also add a timeout to the queue.
Add more functionality trying to get firmware updates working.
Add a cancel firmware update command to master protocol.
Add ability for FC0 to trigger the continuation of a firmware download.