Add debug log functionality.
Add field to each device which effectively controls the logging of all
messages to the debug log api. This defaults to off and can be set to
true only while a device is connected. If the device reconnects to
simchatserver it will not retain the debug flag.
Add ability to set debug through simchatmaster interfaces.